% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cobemo}
\alias{cobemo}
\title{The Copula Benchmark Data (Moderate dependence)}
\format{
A data frame with 9360 rows and 3 variables:
\describe{
  \item{u}{the first uniform scores}
  \item{v}{the second uniform scores}
  \item{copula}{copula families, including "gaussian", "t", "clayton", "gumbel", "frank", "joe", "bb1", "bb6", "bb7", "bb8", "GGEE", "PPPP"}
}
}
\source{
Simulated and used in the paper: Hua, L. (2025), Amortized Neural Inference on Bivariate Tail Dependence and Tail Asymmetry
}
\usage{
cobemo
}
\description{
This benchmark data can be used to compare modeling capacity of different copula models when the model is mis-specified.
}
\examples{
data(cobemo)
head(cobemo)
}
\keyword{datasets}
