% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fasttail}
\alias{fasttail}
\title{Fast inference for bivariate tail dependence and tail asymmetry}
\source{
Based on Hua, L. (2025), Amortized Neural Inference on Bivariate Tail Dependence and Tail Asymmetry
}
\usage{
fasttail(dat, model = "GGEE", random = "quasi")
}
\arguments{
\item{dat}{input of uniform scores, should be n rows and 2 columns; if not uniform scores the data will be converted to uniform scores for further calculation.}

\item{model}{GGEE or PPPP copulas, default is GGEE copula}

\item{random}{Methods for generating random samples used for training NBE, default is quasi}
}
\value{
estimated parameters (alpha, beta > 0) of the copula, and the unified tail dependence parameters (0 < utd_lower, utd_upper < 1, the larger the stronger degree of dependence in the tails)
}
\description{
Fast inference for bivariate tail dependence and tail asymmetry. The first run is a bit slower as it needs to load the model, and after the first run the speed will be ultra fast.
}
\examples{
\dontrun{
dat <- CopulaOne::rGGEE_COP(700, al=1.4, be=0.8)
fasttail(dat)
}
}
\keyword{copula}
\keyword{dependence,}
\keyword{intermediate}
\keyword{tail}
